+++ packages/gamut/src/Form/__tests__/Input.test.tsx
@@ -3,6 +3,24 @@ import { setupRtl } from '@codecademy/gamut-tests';
 
 import { Input } from '../Input';
 
+jest.mock('@codecademy/gamut-icons', () => ({
+  AlertIcon: () => (
+    <svg>
+      <title>Alert Icon</title>
+    </svg>
+  ),
+  CheckCircledIcon: () => (
+    <svg>
+      <title>Check Circled Icon</title>
+    </svg>
+  ),
+  StreakIcon: () => (
+    <svg>
+      <title>Streak Icon </title>
+    </svg>
+  ),
+}));
+
 const renderView = setupRtl(Input, { 'aria-label': 'input' });
 
 describe('Input', () => {